refactor(hydro_lang)!: move `rewrites` module under `builder`
This module will soon go away entirely, since we will no longer rely on the `persist_pullup` rewrite for translating between Flo / batched-DFIR semantics. But in the meantime, we can at least move it from the top-level namespace.
