+++ hydro_lang/src/builder/mod.rs
@@ -17,7 +17,12 @@ use ir::HydroRoot;
 
 #[cfg(feature = "build")]
 #[cfg_attr(docsrs, doc(cfg(feature = "build")))]
+pub mod rewrites;
+
+#[cfg(feature = "build")]
+#[cfg_attr(docsrs, doc(cfg(feature = "build")))]
 pub mod built;
+
 #[cfg(feature = "build")]
 #[cfg_attr(docsrs, doc(cfg(feature = "build")))]
 pub mod compiled;
