+++ hydro_optimize/src/deploy_and_analyze.rs
@@ -4,11 +4,11 @@ use std::time::Duration;
 use hydro_deploy::Deployment;
 use hydro_lang::builder::deploy::DeployResult;
 use hydro_lang::builder::ir::{HydroNode, HydroRoot, deep_clone, traverse_dfir};
+use hydro_lang::builder::rewrites::persist_pullup::persist_pullup;
 use hydro_lang::builder::{FlowBuilder, RewriteIrFlowBuilder};
 use hydro_lang::deploy::HydroDeploy;
 use hydro_lang::deploy::deploy_graph::DeployCrateWrapper;
 use hydro_lang::location::LocationId;
-use hydro_lang::rewrites::persist_pullup::persist_pullup;
 use tokio::sync::mpsc::UnboundedReceiver;
 
 use crate::decouple_analysis::decouple_analysis;
