+++ hydro_optimize/src/partition_node_analysis.rs
@@ -743,11 +743,11 @@ mod tests {
     use std::collections::{BTreeMap, BTreeSet, HashMap};
 
     use hydro_lang::builder::ir::deep_clone;
+    use hydro_lang::builder::rewrites::persist_pullup::persist_pullup;
     use hydro_lang::deploy::HydroDeploy;
     use hydro_lang::live_collections::stream::NoOrder;
     use hydro_lang::location::LocationId;
     use hydro_lang::prelude::*;
-    use hydro_lang::rewrites::persist_pullup::persist_pullup;
     use stageleft::q;
 
     use crate::debug::name_to_id_map;
