+++ hydro_test/src/cluster/compute_pi.rs
@@ -55,9 +55,9 @@ pub fn compute_pi<'a>(
 
 #[cfg(test)]
 mod tests {
+    use hydro_lang::builder::rewrites::persist_pullup;
     use hydro_lang::deploy::HydroDeploy;
     use hydro_lang::location::Location;
-    use hydro_lang::rewrites::persist_pullup;
     use hydro_optimize::decoupler;
 
     struct DecoupledCluster {}
