+++ hydro_test/src/cluster/simple_cluster.rs
@@ -80,9 +80,9 @@ mod tests {
     use std::collections::HashMap;
 
     use hydro_deploy::Deployment;
+    use hydro_lang::builder::rewrites::persist_pullup;
     use hydro_lang::deploy::{DeployCrateWrapper, HydroDeploy};
     use hydro_lang::location::{Location, MemberId};
-    use hydro_lang::rewrites::persist_pullup;
     use hydro_optimize::partitioner::{self, Partitioner};
     use stageleft::q;
